+++ packages/contracts-communication/contracts/events/InterchainClientV1Events.sol
@@ -3,15 +3,15 @@ pragma solidity ^0.8.0;
 
 abstract contract InterchainClientV1Events {
     event ExecutionFeesSet(address executionFees);
-    event LinkedClientSet(uint256 chainId, bytes32 client);
+    event LinkedClientSet(uint64 chainId, bytes32 client);
 
     // TODO: figure out indexing
 
     event InterchainTransactionSent(
         bytes32 indexed transactionId,
-        uint256 indexed dbNonce,
+        uint64 indexed dbNonce,
         uint64 indexed entryIndex,
-        uint256 dstChainId,
+        uint64 dstChainId,
         bytes32 srcSender,
         bytes32 dstReceiver,
         uint256 verificationFee,
@@ -22,14 +22,14 @@ abstract contract InterchainClientV1Events {
 
     event InterchainTransactionReceived(
         bytes32 indexed transactionId,
-        uint256 indexed dbNonce,
+        uint64 indexed dbNonce,
         uint64 indexed entryIndex,
-        uint256 srcChainId,
+        uint64 srcChainId,
         bytes32 srcSender,
         bytes32 dstReceiver
     );
 
     event ExecutionProofWritten(
-        bytes32 indexed transactionId, uint256 indexed dbNonce, uint64 indexed entryIndex, address executor
+        bytes32 indexed transactionId, uint64 indexed dbNonce, uint64 indexed entryIndex, address executor
     );
 }
